New tests, some refactoring, bug fix
The bug fix is the rollback that was already made in the main WASP repo.
I added several tests and refactored code a bit to remove some old stuff
and move some stuff out of main(). Moving the code out of main makes
testing easier and cleans things up a bit.