Bug fix and test updates
I've addressed a bug that was causing filter_remapped_reads.py to fail
for paired end data. I've added a few simple tests to check that the
right reads are output when an incorrectly mapped read pair is either
first, in the middle, or last in the file. I noticed that a couple of
tests had weird behavior and the code didn't make sense to me, so I've
commented those out for now until they can be reviewed. I think some of
these commented out tests were never fully completed.