Bug fix for issue #30
This fixes #30 by adding a check for the end of the remapped bam file and
appropriately exiting. I haven't added any tests for this but it should
be covered by a general test for single end input data for
filter_remapped_reads.py that we should make at some point. Also fixed a
bug in accepting already-sorted bam files for find_intersecting_snps.py.
The command would fail immediately when trying to use the -s option, so
this would have been obvious to anyone and likely hasn't caused any
unknown problems.