Fixed warning about unphased SNPs
Code was checking if a SNP is unphased and if the warn_phase flag is
set in the same conditional.

After unsetting warn_phase, this test would fail, causing additional
unphased SNPs to be converted to missing data.

By creating two separate conditionals, a single warning is printed and
all additional unphased SNPs are parsed normally.